requirements.txt
absl-py==0.15.0
alembic==1.7.5
apache-airflow==1.10.12
apispec==1.3.3
app-store-scraper==0.3.5
argcomplete==1.12.3
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
Babel==2.9.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
cattrs==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
chromedriver-autoinstaller==0.3.1
clang==5.0
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
contextvars==2.4
croniter==0.3.37
cryptography==36.0.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.18.1
email-validator==1.1.3
emoji==1.6.1
entrypoints==0.3
Flask==1.1.4
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
flatbuffers==1.12
funcsigs==1.0.2
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-play-scraper==1.0.2
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httplib2==0.20.2
idna==2.10
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
iso8601==1.0.2
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==2.11.3
joblib==1.1.0
json-merge-patch==0.2
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.1.6
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.27.0
matplotlib==3.3.4
natsort==8.0.2
nest-asyncio==1.5.4
nltk==3.6.7
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.1.5
parso==0.7.1
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prison==0.2.1
prompt-toolkit==3.0.24
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySastrawi==1.2.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
selenium==3.14.1
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.38.2
tabulate==0.8.9
tenacity==4.12.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
thrift==0.15.0
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
trio==0.19.0
trio-websocket==0.9.2
typing==3.7.4.3
typing-extensions==3.7.4.3
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==0.16.1
wrapt==1.12.1
wsproto==1.0.0
WTForms==3.0.0
zipp==3.6.0
zope.deprecation==4.4.0
